+++ vector/src/main/java/im/vector/riotx/core/extensions/Iterable.kt
@@ -38,13 +38,13 @@ inline fun <T, R : Comparable<R>> Iterable<T>.lastMinBy(selector: (T) -> R): T? 
 /**
  * Call each for each item, and between between each items
  */
-inline fun <T> Collection<T>.join(each: (T) -> Unit, between: (T) -> Unit) {
+inline fun <T> Collection<T>.join(each: (Int, T) -> Unit, between: (Int, T) -> Unit) {
     val lastIndex = size - 1
     forEachIndexed { idx, t ->
-        each(t)
+        each(idx, t)
 
         if (idx != lastIndex) {
-            between(t)
+            between(idx, t)
         }
     }
 }