+++ vector/src/main/java/im/vector/riotx/features/invite/InviteUsersToRoomActivity.kt
@@ -30,9 +30,16 @@ import im.vector.riotx.core.di.ScreenComponent
 import im.vector.riotx.core.error.ErrorFormatter
 import im.vector.riotx.core.extensions.addFragment
 import im.vector.riotx.core.extensions.addFragmentToBackstack
+import im.vector.riotx.core.extensions.exhaustive
 import im.vector.riotx.core.platform.SimpleFragmentActivity
 import im.vector.riotx.core.platform.WaitingViewData
+import im.vector.riotx.core.utils.PERMISSIONS_FOR_MEMBERS_SEARCH
+import im.vector.riotx.core.utils.PERMISSION_REQUEST_CODE_READ_CONTACTS
+import im.vector.riotx.core.utils.allGranted
+import im.vector.riotx.core.utils.checkPermissions
 import im.vector.riotx.core.utils.toast
+import im.vector.riotx.features.contactsbook.ContactsBookFragment
+import im.vector.riotx.features.contactsbook.ContactsBookViewModel
 import im.vector.riotx.features.userdirectory.KnownUsersFragment
 import im.vector.riotx.features.userdirectory.KnownUsersFragmentArgs
 import im.vector.riotx.features.userdirectory.UserDirectoryFragment
@@ -53,6 +60,7 @@ class InviteUsersToRoomActivity : SimpleFragmentActivity() {
     private lateinit var sharedActionViewModel: UserDirectorySharedActionViewModel
     @Inject lateinit var userDirectoryViewModelFactory: UserDirectoryViewModel.Factory
     @Inject lateinit var inviteUsersToRoomViewModelFactory: InviteUsersToRoomViewModel.Factory
+    @Inject lateinit var contactsBookViewModelFactory: ContactsBookViewModel.Factory
     @Inject lateinit var errorFormatter: ErrorFormatter
 
     override fun injectWith(injector: ScreenComponent) {
@@ -74,7 +82,8 @@ class InviteUsersToRoomActivity : SimpleFragmentActivity() {
                         UserDirectorySharedAction.Close                 -> finish()
                         UserDirectorySharedAction.GoBack                -> onBackPressed()
                         is UserDirectorySharedAction.OnMenuItemSelected -> onMenuItemSelected(sharedAction)
-                    }
+                        UserDirectorySharedAction.OpenPhoneBook         -> openPhoneBook()
+                    }.exhaustive
                 }
                 .disposeOnDestroy()
         if (isFirstCreation()) {
@@ -92,9 +101,27 @@ class InviteUsersToRoomActivity : SimpleFragmentActivity() {
         viewModel.observeViewEvents { renderInviteEvents(it) }
     }
 
+    private fun openPhoneBook() {
+        // Check permission first
+        if (checkPermissions(PERMISSIONS_FOR_MEMBERS_SEARCH,
+                        this,
+                        PERMISSION_REQUEST_CODE_READ_CONTACTS,
+                        0)) {
+            addFragmentToBackstack(R.id.container, ContactsBookFragment::class.java)
+        }
+    }
+
+    override fun onRequestPermissionsResult(requestCode: Int, permissions: Array<String>, grantResults: IntArray) {
+        if (allGranted(grantResults)) {
+            if (requestCode == PERMISSION_REQUEST_CODE_READ_CONTACTS) {
+                addFragmentToBackstack(R.id.container, ContactsBookFragment::class.java)
+            }
+        }
+    }
+
     private fun onMenuItemSelected(action: UserDirectorySharedAction.OnMenuItemSelected) {
         if (action.itemId == R.id.action_invite_users_to_room_invite) {
-            viewModel.handle(InviteUsersToRoomAction.InviteSelectedUsers(action.selectedUsers))
+            viewModel.handle(InviteUsersToRoomAction.InviteSelectedUsers(action.invitees))
         }
     }